Release assertion failure in Optional<WebCore::SimpleRange>::operator* via CompositeEditCommand::moveParagraphs https://bugs.webkit.org/show_bug.cgi?id=218494
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2020-11-18
Reviewed by Ryosuke Niwa.

Source/WebCore:

This is happening when insert list command is called for a list item containing a body element as a child. When
the tree is iterated looking for the end position, the body element selected as candidate, but a null position
is returned because it's considered to be in a different editing element. This happens because
Node::rootEditableElement() always returns the node itseld for body elements, but it should actually check that
the node is the document body.

Test: editing/inserting/insert-list-with-body-child-crash.html

* dom/Node.cpp:
(WebCore::Node::isRootEditableElement const): Check node is the document body, not just a body element.
(WebCore::Node::rootEditableElement const): Ditto.
* editing/CompositeEditCommand.cpp:
(WebCore::CompositeEditCommand::moveParagraphs): Add an assert to ensure it's not called with a null endOfParagraphToMove.

LayoutTests:

* editing/inserting/insert-list-with-body-child-crash-expected.txt: Added.
* editing/inserting/insert-list-with-body-child-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@269946 268f45cc-cd09-0410-ab3c-d52691b4dbfc