[css-flex] Images as flex items should use the overridingLogicalWidth when defined to compute the logical height https://bugs.webkit.org/show_bug.cgi?id=219195
Reviewed by Manuel Rego Casasnovas.

LayoutTests/imported/w3c:

Replaced 15 FAIL by PASS expectations.

* web-platform-tests/css/css-flexbox/image-as-flexitem-size-001-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-002v-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-005-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-006v-expected.txt:
* web-platform-tests/css/css-flexbox/image-as-flexitem-size-007-expected.txt:

Source/WebCore:

This is the same fix than the one we did in r270073 but for the other axis. RenderReplaced should use the overridingLogicalWidth
whenever defined instead of the specified logical width to compute the logical height using an intrinsic aspect ratio.
The overriding width is set by flex containers that need to stretch/shrink their items. The current code was not considering
this case and thus, the intrinsic (non-stretched) logical height was used to compute the logical width,
meaning that the stretching set by the flexbox container was ignored.

This patch allows WebKit to pass 15 subtests in already existing aspect ratio flexbox tests from WPT.

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::computeReplacedLogicalHeight const): Use the overriding logical width if defined in presence of a valid aspect ratio.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270116 268f45cc-cd09-0410-ab3c-d52691b4dbfc