[iOS][FCR] Add new look for meter element https://bugs.webkit.org/show_bug.cgi?id=219103 <rdar://problem/71549155>
Reviewed by Wenson Hsieh.

Source/WebCore:

Tests: fast/forms/ios/form-control-refresh/meter/border.html
       fast/forms/ios/form-control-refresh/meter/even-less-good-appearance.html
       fast/forms/ios/form-control-refresh/meter/optimal-appearance.html
       fast/forms/ios/form-control-refresh/meter/suboptimal-appearance.html
       fast/forms/ios/form-control-refresh/meter/width-height.html

* html/HTMLMeterElement.cpp:
(WebCore::HTMLMeterElement::createElementRenderer):
* rendering/RenderTheme.cpp:
(WebCore::RenderTheme::supportsMeter const):
* rendering/RenderTheme.h:

Added additional parameter to supportsMeter to control support at runtime.

* rendering/RenderThemeIOS.h:
* rendering/RenderThemeIOS.mm:
(WebCore::RenderThemeIOS::supportsMeter const):

Implement this method so that a native meter is painted on iOS. Note that
a native appearance for meter is currently supported on macOS, but is
unsupported on iOS.

(WebCore::RenderThemeIOS::paintMeter):

Paint a meter element using the new appearance. The color of the filled
portion matches the element's gauge region, and can be one of three
colors. If the style has an RTL direction, flip the filled portion to
start from the right side of the element. This matches the RTL behavior
for progress bars.

* rendering/RenderThemeMac.h:
* rendering/RenderThemeMac.mm:
(WebCore::RenderThemeMac::supportsMeter const):
* rendering/RenderThemeWin.cpp:
(WebCore::RenderThemeWin::supportsMeter const):
* rendering/RenderThemeWin.h:

LayoutTests:

Added tests to verify the stylability of native meters and to
verify that different states are painted with a different appearance.

* fast/forms/ios/form-control-refresh/meter/border-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/meter/border.html: Added.
* fast/forms/ios/form-control-refresh/meter/even-less-good-appearance-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/meter/even-less-good-appearance.html: Added.
* fast/forms/ios/form-control-refresh/meter/optimal-appearance-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/meter/optimal-appearance.html: Added.
* fast/forms/ios/form-control-refresh/meter/suboptimal-appearance-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/meter/suboptimal-appearance.html: Added.
* fast/forms/ios/form-control-refresh/meter/width-height-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/meter/width-height.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270207 268f45cc-cd09-0410-ab3c-d52691b4dbfc