[GTK][WPE] Async scrolling udpates are blocked in the main thread https://bugs.webkit.org/show_bug.cgi?id=219308
Patch by Alejandro G. Castro <alex@igalia.com> on 2020-12-01
Reviewed by Simon Fraser.

Source/WebCore:

We need to avoid this performance improvement of avoiding the
refresh for the scrolling thread when no wheel events were
recently sent because GTK and WPE kinetic scrolling does not use
wheel events to control the animation. We are discussing to change
this for the future and make the code more similar for every port.

No new tests, not sure if we can add tests for this situation. We
detected it manually.

* page/scrolling/ThreadedScrollingTree.cpp:
(WebCore::ThreadedScrollingTree::displayDidRefresh): This function
is also called by the threaded compositor thread in case of GTK
and WPE. We can not use that condition for GTK and WPE, kinetic
animation does not use the wheel events.

Source/WebKit:

We need to use the EventDispatcher from the frameDone callback in
the ThreadedCompositor, for GTK and WPE we receive the frame
information in that thread. We have to notify the scrolling trees
in the scrolling thread even if the main thread is busy or we will
freeze when the main thread is busy. For that we use the API in
the EventDispatcher.

* Shared/CoordinatedGraphics/threadedcompositor/ThreadedCompositor.cpp:
(WebKit::ThreadedCompositor::sceneUpdateFinished): Call the
EventDispatcher to notify the scrolling trees a frame was rendered.
* WebProcess/WebPage/EventDispatcher.cpp: Enable the compilation
of the notify function for all the ports.
* WebProcess/WebPage/EventDispatcher.h: Ditto.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270292 268f45cc-cd09-0410-ab3c-d52691b4dbfc