[GStreamer] Common module should use a dedicated logging category https://bugs.webkit.org/show_bug.cgi?id=219348
Patch by Philippe Normand <pnormand@igalia.com> on 2020-12-01
Reviewed by Xabier Rodriguez-Calvar.

Because this module can be used outside of the MediaPlayer scope, we cannot assume the
MediaPlayer debug category will be initialized in time when GST_DEBUG logs need to be
emitted from GStreamerCommon. So it is safer to introduce a new, standalone, logging
category for this module.

* platform/graphics/gstreamer/GStreamerCommon.cpp:
(WebCore::initializeGStreamer):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270301 268f45cc-cd09-0410-ab3c-d52691b4dbfc