Fix for accessibility layout tests involving ranges in isolated tree mode. https://bugs.webkit.org/show_bug.cgi?id=219436
Reviewed by Chris Fleizach.

Tests:
accessibility/mac/bounds-for-range.html
accessibility/misspelling-range.html

- Implementation of AXIsolatedObject::boundsForRange and misspellingRange.
- Fix for bug in stringForRange that wasn't isolatedCopying the returned
string. Also this method should not dispatch the call to the main thread
since the caller needs to do so because it is passing a SimpleRange as a
parameter.
- Implementation of AXIsolatedObject::lineForPosition. Other methods
that take or return VisiblePositions and VisiblePositionRanges need also
to be implemented.

* accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::stringForRange const):
(WebCore::AXIsolatedObject::boundsForRange const):
(WebCore::AXIsolatedObject::misspellingRange const):
(WebCore::AXIsolatedObject::lineForPosition const):
* accessibility/isolatedtree/AXIsolatedObject.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270352 268f45cc-cd09-0410-ab3c-d52691b4dbfc