MediaSessionHelper::setSharedHelper() hangs when "media in the GPU process" is enabled https://bugs.webkit.org/show_bug.cgi?id=219466 <rdar://problem/71566601>
Reviewed by Chris Dumez.

The WebProcess will attempt to set a RemoteMediaSessionHelper as the sharedHelper, but calling
MediaSessionHelper::setSharedHelper() will cause a MediaSessionHelperIOS to be created, which in
turn tries to talk to AVSystemController.sharedAVSystemController, which fails due to sandbox
restrictions.

Refactor sharedHelperInstance() to not create a MediaSessionHelperIOS by default, which allows
setSharedHelper to assign a new helper without creating one by default.

* platform/audio/ios/MediaSessionHelperIOS.mm:
(sharedHelperInstance):
(MediaSessionHelper::sharedHelper):
(MediaSessionHelper::resetSharedHelper):
(MediaSessionHelper::setSharedHelper):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270394 268f45cc-cd09-0410-ab3c-d52691b4dbfc