Add check to ensure that the size of the result of AXTextMarkerGetBytePtr is equal to the size of TextMarkerData. https://bugs.webkit.org/show_bug.cgi?id=219601
Reviewed by Simon Fraser.

Covered by existing tests.

This is a follow up change to the patch for
https://bugs.webkit.org/show_bug.cgi?id=219491
per Simon Fraser's comment.

* accessibility/mac/AXObjectCacheMac.mm:
(WebCore::getBytesFromAXTextMarker):
Added a check to ensure that the memory block returned by AXTextMarkerGetBytePtr
is the same size as TextMarkerData. This doesn't guaranty that the
underlying binary structures are the same, but it is a sanity check that
may catch flagrant discrepancies.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270523 268f45cc-cd09-0410-ab3c-d52691b4dbfc