REGRESSION(r254841): Possible null pointer reference in BitmapImage::preTransformedNativeImageForCurrentFrame() https://bugs.webkit.org/show_bug.cgi?id=219651
Reviewed by Tim Horton.

preTransformedNativeImageForCurrentFrame() was added in r254841 under the
name nativeImageForCurrentFrameRespectingOrientation() without checking
whether the resulted image of nativeImageForCurrentFrame() is null or not.
If the ImageSource is corrupted, the decoder might return a null
PlatformImagePtr for the current frame. And if this happens most likely
ImageSource::densityCorrectedSize() will return WTF::nullopt also. In this
case, we end up calling GraphicsContext::drawPlatformImage() with a null
PlatformImagePtr.

* platform/graphics/BitmapImage.cpp:
(WebCore::BitmapImage::preTransformedNativeImageForCurrentFrame):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270562 268f45cc-cd09-0410-ab3c-d52691b4dbfc