[Media in GPU Process][MSE] The GPU Process may crash in the destructor of MediaPlayerPrivateMediaSourceAVFObjC https://bugs.webkit.org/show_bug.cgi?id=219622
Reviewed by Eric Carlson.

The purpose of `mediaPlayerRenderingModeChanged()` callback is to notify the media element
that the player's rendering mode is changed, including the case that the layer is destroyed.
So we should not operate on the layer in this function.

With this patch, `RemoteMediaPlayerProxy` forwards the callback from the GPU process to
the Web process. This patch also cleans up `RemoteMediaPlayerProxy.h`.

* GPUProcess/media/RemoteMediaPlayerProxy.cpp:
(WebKit::RemoteMediaPlayerProxy::mediaPlayerRenderingModeChanged):
* GPUProcess/media/RemoteMediaPlayerProxy.h:
* GPUProcess/media/cocoa/RemoteMediaPlayerProxyCocoa.mm:
(WebKit::RemoteMediaPlayerProxy::mediaPlayerRenderingModeChanged): Deleted.
* GPUProcess/media/gstreamer/RemoteMediaPlayerProxyGStreamer.cpp:
(WebKit::RemoteMediaPlayerProxy::mediaPlayerRenderingModeChanged): Deleted.
* GPUProcess/media/win/RemoteMediaPlayerProxyWin.cpp:
(WebKit::RemoteMediaPlayerProxy::mediaPlayerRenderingModeChanged): Deleted.

* WebProcess/GPU/media/MediaPlayerPrivateRemote.cpp:
(WebKit::MediaPlayerPrivateRemote::renderingModeChanged):
* WebProcess/GPU/media/MediaPlayerPrivateRemote.h:
* WebProcess/GPU/media/MediaPlayerPrivateRemote.messages.in:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270563 268f45cc-cd09-0410-ab3c-d52691b4dbfc