[GPUProcess] Video does not resume if the GPU Process crashes while in fullscreen https://bugs.webkit.org/show_bug.cgi?id=219660
Reviewed by Darin Adler.

When the GPU process was crashing while a video was playing in full screen (very common on iOS),
we would attempt to resume playback but video rendering was broken. It looks like this should
be recoverable in some way since changing the orientation of the device makes the video render
again. However, the full screen logic lives in the UIProcess and I could not figure out how to
recover nicely. As a result, this patch simply exits fullscreen when the GPU process crashes.
I have verified that the video renders properly if the user re-enter full screen after a crash.

* platform/graphics/MediaPlayer.cpp:
(WebCore::MediaPlayer::reloadAndResumePlaybackIfNeeded):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270586 268f45cc-cd09-0410-ab3c-d52691b4dbfc