REGRESSION (r269812): Amazon Prime: thumbnail fails to expand properly https://bugs.webkit.org/show_bug.cgi?id=219692 <rdar://problem/71878636>
Reviewed by Simon Fraser.

LayoutTests/imported/w3c:

Correctly fail tests where the "position" property is expected to start a transition
and correctly pass tests some other "position" tests.

* web-platform-tests/css/css-logical/animations/float-interpolation-expected.txt:
* web-platform-tests/css/css-position/animations/position-interpolation-expected.txt:
* web-platform-tests/css/css-transitions/properties-value-003-expected.txt:

Source/WebCore:

We regressed hovering over titles on the main page of Amazon Prime Video since those hover effects
change "position" from "static" to "absolute" and have "transition-property" set to "all".

The CSS Transitions specification defines the following in its "Starting of transitions" section
(https://drafts.csswg.org/css-transitions/#starting):

    "When comparing the before-change style and after-change style for a given property, the property
    values are transitionable if they have an animation type that is neither not animatable nor discrete."

AnimationTimeline::updateCSSTransitionsForStyleableAndProperty(), which implements the specification
text of the "Starting of transitions" section, relies on CSSPropertyAnimation::canPropertyBeInterpolated()
to determine whether a property can be interpolated. This function, in turn, uses the AnimationPropertyWrapperBase
method canInterpolate() to determine support for interpolation.

When adding support for discrete properties, we neglected to make DiscretePropertyWrapper implement this method
and return false, which we correct in this patch.

Fixing this uncovered two issues. First, the WPT test css/css-transitions/properties-value-003.html incorrectly
thinks that the "position" property can be interpolated and start transitions. The WPT issue #26822 deals with
this issue.

Second, text-decoration-color is mistakenly marked as discrete in our implementation. This will be dealt with
in bug 219691.

* animation/CSSPropertyAnimation.cpp:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270589 268f45cc-cd09-0410-ab3c-d52691b4dbfc