[iOS] TestWebKitAPI.GPUProcess.WebProcessTerminationAfterTooManyGPUProcessCrashes is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=219719 <rdar://problem/72158303>
Reviewed by Darin Adler.

We currently have an issue where the audio track gets removed after the GPUProcess
crashes. This is what is causing GPUProcess.WebProcessTerminationAfterTooManyGPUProcessCrashes
to be flaky on iOS. For the same reason, GPUProcess.CrashWhilePlayingVideo was flaky on
iOS until we temporarily disabled the checks in r270328. For now, do the same thing as in r270328
for WebProcessTerminationAfterTooManyGPUProcessCrashes since I have not found the reason why
we lose audio on iOS after a GPU process crash yet and we don't want to keep a flaky test on
the bots.

* TestWebKitAPI/Tests/WebKitCocoa/GPUProcess.mm:
(TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270628 268f45cc-cd09-0410-ab3c-d52691b4dbfc