[GPU Process] Crash when loading drinktrade.com https://bugs.webkit.org/show_bug.cgi?id=219752
Reviewed by Tim Horton.

Fix an overrelease when rendering color emoji fonts by removing an extraneous `adoptCF` around a `CGImageRef`.
This `CGImageRef` is not a retained (+1) object, so we need to make sure that it is retained when tracking it in
the display list's native image cache.

Layout test: fast/text/emoji.html

* platform/graphics/displaylists/DisplayListDrawGlyphsRecorderCoreText.cpp:
(WebCore::DisplayList::DrawGlyphsRecorder::recordDrawImage):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270653 268f45cc-cd09-0410-ab3c-d52691b4dbfc