[Mac] Create a format reader plug-in for WebM https://bugs.webkit.org/show_bug.cgi?id=218908 <rdar://problem/71373264>
Reviewed by Eric Carlson.

Source/WebCore/PAL:

* pal/spi/cocoa/MediaToolboxSPI.h: Defined SPI for MTPluginFormatReader.

Source/WebKit:

Added a new bundle target that builds a MediaToolbox format reader
plug-in and installs it in the WebContent and GPU XPC service bundles.
The plug-in does nothing for now, but in follow-up patches will learn
how to read WebM containers.

* Configurations/MediaFormatReader.xcconfig: Added.
* MediaFormatReaderPlugIn/Info.plist: Added. Defines a CMClassImplementations array
specifying the plug-in factory function and the supported media types / extensions.
* MediaFormatReaderPlugIn/MediaFormatReaderPlugIn.cpp: Added.
* WebKit.xcodeproj/project.pbxproj:

Source/WTF:

* wtf/PlatformHave.h: Defined HAVE_MT_PLUGIN_FORMAT_READER.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270666 268f45cc-cd09-0410-ab3c-d52691b4dbfc