[CMake] Determine correct visibility for linked frameworks https://bugs.webkit.org/show_bug.cgi?id=210366
Reviewed by Michael Catanzaro.

.:

WebKit has some frameworks that are compiled statically into a larger shared library. For
example Apple compiles bmalloc and WTF into JavaScriptCore. This means the JavaScriptCore
framework also exposes the symbols for those two libraries and that targets depending
on WTF or bmalloc should link to JavaScriptCore and never to WTF or bmalloc directly.

The CMake code allows each port to decide the type of library for a WEBKIT_FRAMEWORK.
This means there are many different combinations that it needs to support. To ensure that
everything links properly some bookkeeping is required.

WEBKIT_FRAMEWORK now uses ${target}_FRAMEWORKS to determine how to link the frameworks
it depends on. First the macro determines if the framework is a SHARED library or is linked
into a shared library, and if so it is linked as PUBLIC. If the framework itself is not a
SHARED library, then it links other STATIC/OBJECT framework as PUBLIC. If the framework
itself is SHARED then any framework that is a STATIC/OBJECT library is linked as PRIVATE
to stop propagation. The linked framework's header directories are added to the
WebKit::${framework} target to allow the framework to present as that framework. A global
property is set to inform consumers that they should link to this different framework
to get the requested framework.

This ensures that all ports can ship with hidden visibility turned on which is a precursor
to a port turning on link time optimization (LTO).

* Source/cmake/WebKitMacros.cmake:

Source/JavaScriptCore:

Set JavaScriptCore_FRAMEWORKS to determine correct linkage for the library. Remove
explicit setting of STATICALLY_LINKED_WITH_${framework} and $<TARGET_OBJECTS:${framework}>
by ports.

Move the add_subdirectory of shell to the end of the CMakeLists.txt so its after the
WEBKIT_FRAMEWORK call. This ensures that the frameworks linked into JavaScriptCore are
known when creating the executables in that directory.

* CMakeLists.txt:
* PlatformGTK.cmake:
* PlatformJSCOnly.cmake:
* PlatformMac.cmake:
* PlatformPlayStation.cmake:
* shell/CMakeLists.txt:

Source/WebCore:

Set WebCore_FRAMEWORKS and WebCoreTestSupport_FRAMEWORKS to determine correct linkage
for the libraries. Remove explicit setting of STATICALLY_LINKED_WITH_PAL.

For the GTK port JavaScriptCore is compiled as a SHARED library with WTF and bmalloc
compiled into it. WEBKIT_FRAMEWORK now correctly prevents WTF and bmalloc being linked
into targets that also link JavaScriptCore. However GTK port can't be compiled with
hidden visibility turned on currently to properly export symbols so the libraries are
added back to the list for WebCore to keep the old, albiet wrong, behavior.

* CMakeLists.txt:
* PlatformGTK.cmake:

Source/WebCore/PAL:

Set PAL_FRAMEWORKS to determine correct linkage for the library.

* pal/CMakeLists.txt:

Source/WebDriver:

Conditionally add bmalloc to the list of WebDriver_FRAMEWORKS.

* CMakeLists.txt:

Source/WebKit:

Set WebKit_FRAMEWORKS to determine correct linkage for the library. Remove
explicit setting of STATICALLY_LINKED_WITH_${framework} and $<TARGET_OBJECTS:${framework}>
by ports.

* CMakeLists.txt:
* PlatformFTW.cmake:
* PlatformPlayStation.cmake:
* PlatformWPE.cmake:
* PlatformWin.cmake:

Source/WebKitLegacy:

Remove explicit setting of STATICALLY_LINKED_WITH_${framework}.

* PlatformFTW.cmake:
* PlatformWin.cmake:

Source/WTF:

Set WTF_FRAMEWORKS to determine correct linkage for the library. Remove
explicit setting of STATICALLY_LINKED_WITH_${framework} by ports.

* wtf/CMakeLists.txt:
* wtf/PlatformPlayStation.cmake:

Tools:

Remove explicit setting of STATICALLY_LINKED_WITH_${framework}.

* TestWebKitAPI/PlatformWin.cmake:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270690 268f45cc-cd09-0410-ab3c-d52691b4dbfc