[iOS][FCR] Add new look for select elements https://bugs.webkit.org/show_bug.cgi?id=219507 <rdar://problem/71951874>
Reviewed by Wenson Hsieh.

Source/WebCore:

Implement the new in-page appearance for <select> and <select multiple>.
The new appearance/behavior of the select popup will be added separately.

Tests: fast/forms/ios/form-control-refresh/select/background-color.html
       fast/forms/ios/form-control-refresh/select/border-radius.html
       fast/forms/ios/form-control-refresh/select/border.html
       fast/forms/ios/form-control-refresh/select/multiple-appearance.html
       fast/forms/ios/form-control-refresh/select/rtl-appearance.html
       fast/forms/ios/form-control-refresh/select/text-color.html
       fast/forms/ios/form-control-refresh/select/width-height.html

* css/formControlsIOS.css:
(select):
* rendering/RenderMenuList.cpp:
(WebCore::RenderMenuList::adjustInnerStyle):
(RenderMenuList::layout):

Override the layout method so that the border radius can be adjusted to
enforce the pill-shaped appearance. Ideally, we should not be adjusting
styles during layout. However, for a pill-shaped appearance, the
horizontal border radius is dependent on the computed height of the box.
This means that the border radius cannot be declared prior to layout,
since CSS only allows the horizontal border radius to be dependent on the
computed width of the box.

Ignoring the style's border radius and forcing a pill-shaped appearance at
paint time is not an option, since focus rings and tap highlights will not
use the correct border radius. Consequently, we need to adjust the border
radius after layout.

Note that similar adjustments to style after layout are made in RenderSliderThumb,
RenderButton, and RenderTextControlSingleLine.

* rendering/RenderMenuList.h:
* rendering/RenderTheme.h:
(WebCore::RenderTheme::popupInternalPaddingBox const):

Added a new parameter, so that the padding can be controlled at runtime.

* rendering/RenderThemeAdwaita.cpp:
(WebCore::RenderThemeAdwaita::popupInternalPaddingBox const):
* rendering/RenderThemeAdwaita.h:
* rendering/RenderThemeIOS.h:
* rendering/RenderThemeIOS.mm:
(WebCore::RenderThemeIOS::popupInternalPaddingBox const):

Use a 1em padding, so that the button scales with the font size.

(WebCore::canAdjustBorderRadiusForAppearance):

Enforce a pill-like shape, only if the author has not specified a
custom border radius.

(WebCore::RenderThemeIOS::paintMenuListButtonDecorations):
(WebCore::RenderThemeIOS::paintMenuListButtonDecorationsWithFormControlRefresh):

Paint the appropriate glyph for <select> (a two-headed arrow) and
<select multiple> (ellipses). Both glyphs were obtained from
SF Symbols. The glyph is painted on the right by default, and on
the left if the element's direction is RTL.

* rendering/RenderThemeMac.h:
* rendering/RenderThemeMac.mm:
(WebCore::RenderThemeMac::popupInternalPaddingBox const):

LayoutTests:

Added tests to verify the stylability of select elements with the new appearance.

* fast/forms/ios/form-control-refresh/select/background-color-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/select/background-color.html: Added.
* fast/forms/ios/form-control-refresh/select/border-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/select/border-radius-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/select/border-radius.html: Added.
* fast/forms/ios/form-control-refresh/select/border.html: Added.
* fast/forms/ios/form-control-refresh/select/multiple-appearance-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/select/multiple-appearance.html: Added.
* fast/forms/ios/form-control-refresh/select/rtl-appearance-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/select/rtl-appearance.html: Added.
* fast/forms/ios/form-control-refresh/select/text-color-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/select/text-color.html: Added.
* fast/forms/ios/form-control-refresh/select/width-height-expected-mismatch.html: Added.
* fast/forms/ios/form-control-refresh/select/width-height.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270713 268f45cc-cd09-0410-ab3c-d52691b4dbfc